Moved creation of strongly typed param value from GlobalRunSpec to RunnableProjectGenerator
