+++ src/Microsoft.TemplateEngine.Orchestrator.RunnableProjects/RunnableProjectGenerator.cs
@@ -9,6 +9,7 @@ using Microsoft.TemplateEngine.Core.Contracts;
 using Microsoft.TemplateEngine.Utils;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using System.Globalization;
 
 namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects
 {
@@ -114,6 +115,179 @@ namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects
             return template != null;
         }
 
+        //
+        // Converts the raw, string version of a parameter to a strongly typed value.
+        // If the param has a datatype specified, use that. Otherwise attempt to infer the type.
+        // Throws a TemplateParamException if the conversion fails for any reason.
+        //
+        public object ConvertVariableValueToType(ITemplateParameter parameter, string untypedValue)
+        {
+            if (untypedValue == null)
+            {
+                throw new TemplateParamException("Parameter value is null", parameter.Name, null, parameter.DataType);
+            }
+
+            if (!string.IsNullOrEmpty(parameter.DataType))
+            {
+                object convertedValue = DataTypeSpecifiedConvertLiteral(parameter, untypedValue);
+
+                if (convertedValue == null)
+                {
+                    throw new TemplateParamException("Parameter value could not be converted", parameter.Name, untypedValue, parameter.DataType);
+                }
+
+                return convertedValue;
+            }
+            else
+            {
+                return InferTypeAndConvertLiteral(untypedValue);
+            }
+        }
+
+        // For explicitly data-typed variables, attempt to convert the variable value to the specified type.
+        // Data type names:
+        //     - choice
+        //     - bool
+        //     - float
+        //     - int
+        //     - hex
+        //     - text
+        // The data type names are case insensitive.
+        //
+        // Returns the converted value if it can be converted, throw otherwise
+        private static object DataTypeSpecifiedConvertLiteral(ITemplateParameter param, string literal)
+        {
+            if (string.Equals(param.DataType, "bool", StringComparison.OrdinalIgnoreCase))
+            {
+                if (string.Equals(literal, "true", StringComparison.OrdinalIgnoreCase))
+                {
+                    return true;
+                }
+                else if (string.Equals(literal, "false", StringComparison.OrdinalIgnoreCase))
+                {
+                    return false;
+                }
+                else
+                {
+                    // Note: if the literal is ever null, it is probably due to a problem in TemplateCreator.Instantiate()
+                    // which takes care of making null bool -> true as appropriate.
+                    // This else can also happen if there is a value but it can't be converted.
+                    throw new TemplateParamException("Value is not a bool", param.Name, literal, param.DataType);
+                }
+            }
+            else if (string.Equals(param.DataType, "choice", StringComparison.OrdinalIgnoreCase))
+            {
+                if ((literal != null) && param.Choices.Contains(literal))
+                {
+                    return literal;
+                }
+                else
+                {
+                    string conversionErrorMessage = string.Format("Choice is invalid. Valid choices are: [{0}]", string.Join(",", param.Choices));
+                    throw new TemplateParamException(conversionErrorMessage, param.Name, literal, param.DataType);
+                }
+            }
+            else if (string.Equals(param.DataType, "float", StringComparison.OrdinalIgnoreCase))
+            {
+                double convertedFloat;
+                if (double.TryParse(literal, out convertedFloat))
+                {
+                    return convertedFloat;
+                }
+                else
+                {
+                    throw new TemplateParamException("Value is not a float", param.Name, literal, param.DataType);
+                }
+            }
+            else if (string.Equals(param.DataType, "int", StringComparison.OrdinalIgnoreCase))
+            {
+                long convertedInt;
+                if (long.TryParse(literal, out convertedInt))
+                {
+                    return convertedInt;
+                }
+                else
+                {
+                    throw new TemplateParamException("Value is not an int", param.Name, literal, param.DataType);
+                }
+            }
+            else if (string.Equals(param.DataType, "hex", StringComparison.OrdinalIgnoreCase))
+            {
+                long convertedHex;
+                if (long.TryParse(literal.Substring(2), NumberStyles.HexNumber, CultureInfo.InvariantCulture, out convertedHex))
+                {
+                    return convertedHex;
+                }
+                else
+                {
+                    throw new TemplateParamException("Value is not hex format", param.Name, literal, param.DataType);
+                }
+            }
+            else if (string.Equals(param.DataType, "text", StringComparison.OrdinalIgnoreCase))
+            {   // "text" is a valid data type, but doesn't need any special handling.
+                return literal;
+            }
+            else
+            {
+                string customMessage = string.Format("Param name = [{0}] had unknown data type = [{1}]", param.Name, param.DataType);
+                throw new TemplateParamException(customMessage, param.Name, literal, param.DataType);
+            }
+        }
+
+        private static object InferTypeAndConvertLiteral(string literal)
+        {
+            if (literal == null)
+            {
+                return null;
+            }
+
+            if (!literal.Contains("\""))
+            {
+                if (string.Equals(literal, "true", StringComparison.OrdinalIgnoreCase))
+                {
+                    return true;
+                }
+
+                if (string.Equals(literal, "false", StringComparison.OrdinalIgnoreCase))
+                {
+                    return false;
+                }
+
+                if (string.Equals(literal, "null", StringComparison.OrdinalIgnoreCase))
+                {
+                    return null;
+                }
+
+                double literalDouble;
+                if (literal.Contains(".") && double.TryParse(literal, out literalDouble))
+                {
+                    return literalDouble;
+                }
+
+                long literalLong;
+                if (long.TryParse(literal, out literalLong))
+                {
+                    return literalLong;
+                }
+
+                if (literal.StartsWith("0x", StringComparison.OrdinalIgnoreCase)
+                    && long.TryParse(literal.Substring(2), NumberStyles.HexNumber, CultureInfo.InvariantCulture, out literalLong))
+                {
+                    return literalLong;
+                }
+
+                if (string.Equals("null", literal, StringComparison.OrdinalIgnoreCase))
+                {
+                    return null;
+                }
+
+                return literal;
+            }
+
+            return literal.Substring(1, literal.Length - 2);
+        }
+
+
         internal class ParameterSet : IParameterSet
         {
             private readonly IDictionary<string, ITemplateParameter> _parameters = new Dictionary<string, ITemplateParameter>(StringComparer.OrdinalIgnoreCase);
@@ -129,7 +303,7 @@ namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects
 
             public IEnumerable<ITemplateParameter> ParameterDefinitions => _parameters.Values;
 
-            public IDictionary<ITemplateParameter, string> ResolvedValues { get; } = new Dictionary<ITemplateParameter, string>();
+            public IDictionary<ITemplateParameter, object> ResolvedValues { get; } = new Dictionary<ITemplateParameter, object>();
 
             public IEnumerable<string> RequiredBrokerCapabilities => Enumerable.Empty<string>();
